requirements.txt
Django
psycopg2-binary
djangorestframework
django-filter
python-decouple
drf-yasg